decomp: r2 lane-3 shared-header drift fixes (+16 strict, 0 regressions)
Four retail-witnessed shared-header/source fixes from the offset-drift
sweep round-2 recon dossiers:

1. LabelNumberTicker.h: delete Wii-dev-only `u32 unk74` mid-class member
   (retail has zero gap before 8-aligned mTimer; sizeof 0x1e8->0x1e0).
   +8: SyncProperty, SetDesiredValue, Save, Poll, fn_828029F8/2A3C/2E24/2E54.

2. InlineHelp.h/.cpp: delete DC3-only `ResourceDirPtr<ObjectDir>
   mResourceDir` (retail uses inherited UIComponent::mResourceDir;
   vbase anchor 0x194->0x184, sizeof 0x1c8->0x1b8). OldResourcePreload
   reworked to the inline SetName equivalent (unmatched fn, regression-free).
   +5: AppInlineHelp Handle, ctor, UpdateIconTypes, fn_825E7230/7274.

3. obj/Utl.h -> obj/Utl.cpp: SafeName de-inlined (retail witness
   fn_82733060 out-of-line, every cross-TU user calls it); wired
   system/obj/Utl.cpp as NonMatching. Plus (float) casts on the fmod
   results in BandCharacter::UpdateOverlay (retail frsp rounds both
   MakeString args to single). +1: UpdateOverlay 85.2->100.

4. math/Utl.h: delete DC3-era ClampEq/MinEq/MaxEq<float> fsel
   specializations (retail compiled branchy generics; witness
   fn 0x82428440 UtilDrawPlane region). +2: CharLookAt::SyncLimits,
   CharWeightSetter::Poll; UtilDrawPlane 93.1->99.3 fuzzy.

Gate: composed A/B run1==run2 (measures.matched_functions 14520,
baseline 14504); measure_delta NET +16 / 0 strict / 0 fuzzy regressions
(eps 1.0); icf_alias_check HONEST (15/15 real-bodied).

+++ src/system/bandobj/BandCharacter.cpp
@@ -641,8 +641,8 @@ void BandCharacter::UpdateOverlay() {
                               << CharClip::BeatAlignString(firstaddplaying->mPlayFlags);
                     *mOverlay << MakeString(
                         " %.2f %.2f",
-                        std::fmod(TheTaskMgr.Beat(), 1.0f),
-                        std::fmod(firstaddplaying->mBeat, 1.0f)
+                        (float)std::fmod(TheTaskMgr.Beat(), 1.0f),
+                        (float)std::fmod(firstaddplaying->mBeat, 1.0f)
                     );
                 } else {
                     *mOverlay << " "
@@ -651,8 +651,8 @@ void BandCharacter::UpdateOverlay() {
                               << CharClip::BeatAlignString(firstplaying->mPlayFlags);
                     *mOverlay << MakeString(
                         " %.2f %.2f",
-                        std::fmod(TheTaskMgr.Beat(), 1.0f),
-                        std::fmod(firstplaying->mBeat, 1.0f)
+                        (float)std::fmod(TheTaskMgr.Beat(), 1.0f),
+                        (float)std::fmod(firstplaying->mBeat, 1.0f)
                     );
                 }
             } else {
@@ -661,8 +661,8 @@ void BandCharacter::UpdateOverlay() {
                 *mOverlay << " " << CharClip::BeatAlignString(firstplaying->mPlayFlags);
                 *mOverlay << MakeString(
                     " %.2f %.2f",
-                    std::fmod(TheTaskMgr.Beat(), 1.0f),
-                    std::fmod(firstplaying->mBeat, 1.0f)
+                    (float)std::fmod(TheTaskMgr.Beat(), 1.0f),
+                    (float)std::fmod(firstplaying->mBeat, 1.0f)
                 );
             }
         }
